Add self describing meta data to file callback
This commit will add an extra property fileDescriptor to the onFileData callback when we are getting data for session. This property basically describes the data that is returned in this data file. How many there are, what service group it's from, what object type they are and what service the data was retreived from.