Add support for ongoing private share
This commit will add a new feature to the SDK that allows third party
developers to start a private share with their users and retain an
access token which they will be able to use in the future to access user
data in the future.

Also updated in this commit is the copyright year at the header of every
file.