Continue polling on library pending
This commit updates the getSessionData function to continue polling when we receive a pending status from server. Before this release, when a third party calls getSessionData and if the status is pending, we would simply quit the call.

With the introduction of CA MQ (Consent Access Message Queue) on the server, there might be a slight delay when the digi.me client triggers the data query to when the library status returns running. We therefore need to update the SDK so that getSessionData continues polling even when receiving a pending status.